requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
billiard
boto3==1.26.48
botocore==1.29.48
bs4==0.0.1
cached-property==1.5.2
celery>=5.3,<6
celery-progress==0.1.3
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-default-group==1.2.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.0
datapackage==1.15.2
decorator==5.1.1
deepdiff==6.2.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.1
django-bootstrap-breadcrumbs==0.9.1
django-bootstrap-modal-forms==1.4.1
django-celery-beat
django-celery-results
django-cors-headers==2.4.0
django-datatables-view==1.17.0
django-debug-toolbar==4.2.0
django-encrypted-model-fields==0.6.5
django-extensions==3.2.1
django-filter==22.1
django-fontawesome==1.0
django-geojson==4.0.0
django-guardian==2.3.0
django-health-check
django-leaflet==0.28.3
django-maintenance-mode
django-mathfilters==0.4.0
django-multiselectfield==0.1.12
django-ranged-response==0.2.0
django-redis==5.4.0
django-resized==0.3.11
django-shapeshifter==18.9.23
django-storages
django-tables2==2.3.4
django-templated-mail==1.1.1
django-timezone-field==5.0
django-tinymce==3.3.0
django-user-agents==0.4.0
django-webpack-loader==0.6.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.28.0
docutils==0.19
edtf==4.0.1
elasticsearch8==8.9.0
et-xmlfile==1.1.0
flower
fontawesome==5.10.1.post1
frictionless==3.31.0
fsspec==2024.6.1
geojson==3.1.0
geopy==2.4.1
goodtables==2.5.4
greenlet==2.0.2
gunicorn==21.2.0
humanize==4.7.0
idna==3.4
ijson==3.3.0
isodate
jmespath==1.0.1
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
kombu>=5.3,<6
linear-tsv==1.1.0
mozilla-django-oidc==4.0.1
nameparser==1.1.3
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.9
ordered-set==4.1.0
pandas==1.4.1
parameterized==0.7.4
petl==1.7.12
Pillow==9.4.0
pip-search==0.0.12
prometheus-client==0.17.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
pyld==2.0.4
PyJWT==2.6.0
pyproj==3.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
rdflib
rdflib-jsonld
redis==4.5.2
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==12.6.0
s3transfer==0.6.0
scikit-learn>=1.0
sendgrid==6.11.0
sentry-sdk[django]
shapely==2.0.6
shellingham==1.5.0.post1
simpleeval==0.9.12
simplejson==3.18.1
six==1.16.0
slack-sdk
soupsieve==2.4
SQLAlchemy==1.4.46
sqlparse==0.4.3
statistics==1.0.3.5
stringcase==1.2.0
tableschema==1.20.2
tabulator==1.53.5
text-unidecode==1.3
tornado==6.3.2
typer==0.7.0
ua-parser==0.18.0
unicodecsv==0.14.1
urllib3==1.26.14
user-agents==2.2.0
validators==0.18.2
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
slack_sdk~=3.32.0
python-dotenv~=1.0.1
Jinja2~=3.1.4
ratelimit==2.2.1